Move test from EmptyConstructorSpec to EmptyDefaultConstructorSpec
